Fix issues with search books
Properly check each case, book is available if there's at least one copy without borrow and order ID.
Book is orderable if it has at least one copy without an order ID and is being borrowed.
Book is unavailable if it doesn't meet any of these conditions.

Possible bug: if only return_date is defined, this will result in a bug, but this is invalid data.